data_abstractor/requirements.txt
kafka-python==2.0.2
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.13.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
click==8.1.3
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
fqdn==1.5.1
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.11.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.2
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyter_server==2.0.6
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.2
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pylint==2.15.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.19.3
scipy==1.10.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
StrEnum==0.4.9
terminado==0.17.1
tifffile==2022.10.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.8.1
typing_extensions==4.4.0
uri-template==1.2.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
zipp==3.11.0

